+++ CHANGELOG.md
@@ -9,13 +9,16 @@
 ### Changed
 
 - Stop shadowing Gradle plugin dependencies, manually repack only JPS Module
+- Load the `com.intellij` module by default when creating the IntelliJ Platform dependency
 
 ### Fixed
 
 - Performance improvement: memoize the `JavaRuntimePathResolver` resolution.
 - Performance improvement: do not resolve JetBrains IDEs/Android Studio release URLs in the configuration phase.
 - Performance improvement: cache the IntelliJ Platform instances parsed with the Plugin Verifier libraries
-- Tests classpath fixes: do not blindly include all plugin's `lib/**/*.jar`
+- Tests classpath: do not blindly include all plugin's `lib/**/*.jar`
+- Tests classpath: do not load all bundled plugins and modules into the classpath
+- Classpath: correctly resolve all necessary modules for bundled plugins and modules
 
 ## [2.3.0] - 2025-02-27